Can't change "Before Commit" settings without actual commit
Relates to: IDEA-178045
Subtask of: IJP-25
 Some "Before Commit" settings are enabled by default ("Perform code
 analysis" and "Check TODO").
I don't use them. But I realize that after pressing the "Commit" button.
 Every time.
To get rid of these prompts I have to remember to uncheck the options
before the next commit.

You should either provide an option to disable corresponding setting
right on the prompt dialog (similar to "Do not show this again").
Or save the settings when "Commit" dialog is closed with Cancel.
Or add corresponding options to File | Settings | Version Control |
Commit Dialog.
Or all of the above.

P.S. I rarely change these options once I get them setup for my liking.
I guess that's the case for most.
So it might be reasonable to move them to File | Settings | Version
Control | Commit Dialog for good.